# apps/api/openapi/v1.yaml
...
openapi: "3.1.0"
info:
  title: EasyA2 API — Foundation + Org Structure + Identity Advanced + Process Modeling
  version: "1.4.0"
  description: |
    Foundation (MOD-000) + Organizational Structure (MOD-003) + Identity Advanced (MOD-004) + Process Modeling (MOD-005) API.
    All errors follow RFC 9457 Problem Details.
servers:
  - url: /api/v1
...
paths:
...
  /auth/login:
    post:
      operationId: auth_login
      tags: [auth]
      summary: Native login with email and password
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful (or MFA required)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/LoginResponse"
                  - $ref: "#/components/schemas/LoginMfaResponse"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "429":
          $ref: "#/components/responses/ProblemDetails429"
  /auth/logout:
    post:
      operationId: auth_logout
      tags: [auth]
      summary: Logout — revokes current session
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Session revoked
        "401":
          $ref: "#/components/responses/ProblemDetails401"
  /auth/refresh:
    post:
      operationId: auth_refresh
      tags: [auth]
      summary: Refresh access token with rotation
      responses:
        "200":
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
  /auth/me:
    get:
      operationId: auth_get_profile
      tags: [auth]
      summary: Get authenticated user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
    patch:
      operationId: auth_update_profile
      tags: [auth]
      summary: Update own profile
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
  /auth/change-password:
    post:
      operationId: auth_change_password
      tags: [auth]
      summary: Change password (requires current password)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessage"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
  /auth/forgot-password:
    post:
      operationId: auth_forgot_password
      tags: [auth]
      summary: Request password reset email
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Always returns success (BR-001)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessage"
  /auth/reset-password:
    post:
      operationId: auth_reset_password
      tags: [auth]
      summary: Reset password with token
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessage"
        "422":
          $ref: "#/components/responses/ProblemDetails422"
  /auth/sessions:
    get:
      operationId: auth_list_sessions
      tags: [auth]
      summary: List active sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SessionItem"
    delete:
      operationId: auth_revoke_all_sessions
      tags: [auth]
      summary: Revoke all sessions (kill-switch)
      security:
        - bearerAuth: []
      responses:
        "204":
          description: All sessions revoked
  /auth/sessions/{id}:
    delete:
      operationId: auth_revoke_session
      tags: [auth]
      summary: Revoke a specific session
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Session revoked

  # ---------- Users ---------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema:
        type: string
        format: uuid
      description: Idempotency key to prevent duplicate mutations (BR-012)
...
  responses:
    ProblemDetails400:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ProblemDetails401:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    ProblemDetails422:
      description: Unprocessable Entity
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ProblemDetails429:
      description: Too Many Requests
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        extensions:
          type: object
          properties:
            correlationId: { type: string }
    GenericMessage:
      type: object
      properties:
        message: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 255 }
        password: { type: string, minLength: 1 }
        remember_me: { type: boolean, default: false }
        device_fp: { type: string, maxLength: 500 }
    LoginResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
            full_name: { type: string }
            status: { type: string }
    LoginMfaResponse:
      type: object
      properties:
        mfa_required: { type: boolean, enum: [true] }
        temp_token: { type: string }
        expires_in: { type: integer, enum: [300] }
    ProfileResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        codigo: { type: string }
        full_name: { type: string }
        avatar_url: { type: string, nullable: true }
        status: { type: string }
        active_tenant_id: { type: string, format: uuid, nullable: true }
        scopes: { type: array, items: { type: string } }
    UpdateProfileRequest:
      type: object
      properties:
        full_name: { type: string, minLength: 1, maxLength: 255 }
        avatar_url: { type: string, format: uri, nullable: true }
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string, minLength: 1 }
        new_password: { type: string, minLength: 8, maxLength: 128 }
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 255 }
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token: { type: string, format: uuid }
        new_password: { type: string, minLength: 8, maxLength: 128 }
    SessionItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        device_fp: { type: string, nullable: true }
        remember_me: { type: boolean }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        is_current: { type: boolean }
...